Corregidos errores en pruebas y validadas entradas de usuario

+++ src/controller/ControladorUsuarios.py
@@ -1,10 +1,7 @@
+import psycopg2
 import sys
 sys.path.append("src")
-import psycopg2
-import compressor.compressorlogic as compressor
-from model.Usuario import DuplicateEntryError
-from model.Usuario import EntryNotFoundError
-# Configuración desde el módulo de ajustes de seguridad.
+from model.Usuario import UserInput, UserOutput, DuplicateEntryError, EntryNotFoundError, DataValidationError
 import controller.SecretConfig as st
 
 class UserData:
@@ -12,18 +9,11 @@ class UserData:
 
     @staticmethod
     def get_connection():
-        """ 
-        Establece y retorna una conexión a la base de datos utilizando los ajustes especificados en la configuración de seguridad.
-        """
         return psycopg2.connect(database=st.PGDATABASE, user=st.PGUSER, password=st.PGPASSWORD, host=st.PGHOST, port=st.PGPORT)
 
     @staticmethod
     def create_table():
-        """
-        Crea la tabla de usuarios en la base de datos si no existe.
-        Define la estructura de la tabla con columnas para cédula, nombre, teléfono, correo electrónico, tipo de evento,
-        texto original y procesado, y una marca de tiempo por defecto.
-        """
+        """ Crea la tabla de usuarios en la base de datos si no existe. """
         with UserData.get_connection() as conn:
             with conn.cursor() as cursor:
                 cursor.execute("""
@@ -32,78 +22,75 @@ class UserData:
                         Nombre VARCHAR(100) NOT NULL,
                         Telefono BIGINT,
                         CorreoElectronico VARCHAR(100),
-                        tipoEvento VARCHAR(10),
+                        TipoEvento VARCHAR(10),
                         TextoOriginal TEXT,
                         TextoProcesado TEXT,
                         FechaHora TIMESTAMP DEFAULT CURRENT_TIMESTAMP
                     );
                 """)
-                conn.commit()  # Confirma la operación en la base de datos.
+                conn.commit()
 
     @staticmethod
     def drop_table():
-        """
-        Elimina la tabla de usuarios de la base de datos.
-        Este método es útil para limpiar durante pruebas o cuando se necesita reconstruir la tabla.
-        """
         with UserData.get_connection() as conn:
             with conn.cursor() as cursor:
                 cursor.execute("DROP TABLE IF EXISTS Usuarios;")
                 conn.commit()
+                
 
     @staticmethod
     def insert_user(cedula, nombre, telefono, correo, tipo_evento, texto_original, texto_procesado):
-        """
-        Inserta un nuevo usuario en la base de datos. Si se produce un error de integridad (por ejemplo, duplicado de cédula),
-        se revierte la transacción y se lanza una excepción personalizada.
-        """
+        """ Inserta un nuevo usuario en la base de datos tras validar la entrada. """
+        user = UserInput(nombre, cedula, telefono, correo, texto_original, tipo_evento)
+        user.validate()
+        UserInput.check_primary_key(cedula, UserData.user_exists)
         with UserData.get_connection() as conn:
             with conn.cursor() as cursor:
-                try:
-                    cursor.execute("""
-                        INSERT INTO Usuarios (Cedula, Nombre, Telefono, CorreoElectronico, tipoEvento, TextoOriginal, TextoProcesado)
-                        VALUES (%s, %s, %s, %s, %s, %s, %s);
-                    """, (cedula, nombre, telefono, correo, tipo_evento, texto_original, texto_procesado))
-                    conn.commit()
-                except psycopg2.IntegrityError:
-                    conn.rollback()  # Revoca la transacción en caso de error.
-                    raise DuplicateEntryError(f"Ya existe un usuario con la cédula {cedula}.")
+                cursor.execute("""
+                    INSERT INTO Usuarios (Cedula, Nombre, Telefono, CorreoElectronico, TipoEvento, TextoOriginal, TextoProcesado)
+                    VALUES (%s, %s, %s, %s, %s, %s, %s);
+                """, (cedula, nombre, telefono, correo, tipo_evento, texto_original, texto_procesado))
+                conn.commit()
 
     @staticmethod
     def delete_user(cedula):
-        """
-        Elimina un usuario de la base de datos basado en la cédula proporcionada.
-        """
+        """ Elimina un usuario de la base de datos basado en la cédula proporcionada. """
+        if not isinstance(cedula, int):
+            raise DataValidationError("ID inválido.")
         with UserData.get_connection() as conn:
             with conn.cursor() as cursor:
                 cursor.execute("DELETE FROM Usuarios WHERE Cedula = %s;", (cedula,))
+                affected_rows = cursor.rowcount
                 conn.commit()
+        UserOutput.validate_user_found(affected_rows != 0, "eliminar")
 
     @staticmethod
     def update_user(cedula, key, value):
-        """
-        Actualiza la información de un usuario en la base de datos. Si no se encuentra un usuario con la cédula proporcionada,
-        se lanza una excepción de EntryNotFoundError.
-        """
+        """ Actualiza la información de un usuario en la base de datos. """
+        if not isinstance(cedula, int):
+            raise DataValidationError("ID inválido.")
         with UserData.get_connection() as conn:
             with conn.cursor() as cursor:
                 cursor.execute(f"UPDATE Usuarios SET {key} = %s WHERE Cedula = %s;", (value, cedula))
-                if cursor.rowcount == 0:  # Verifica si alguna fila fue actualizada.
-                    conn.rollback()  # Revoca si no hay filas actualizadas.
-                    raise EntryNotFoundError(f"No se encontró un usuario con cédula {cedula} para actualizar.")
+                affected_rows = cursor.rowcount
                 conn.commit()
+        UserOutput.validate_user_found(affected_rows != 0, "actualizar")
 
     @staticmethod
     def query_user(cedula):
-        """
-        Consulta la información de un usuario por su cédula y retorna los datos si están disponibles.
-        Si no se encuentra el usuario, se lanza una excepción de EntryNotFoundError.
-        """
+        """ Consulta la información de un usuario por su cédula. """
+        if not isinstance(cedula, int):
+            raise DataValidationError("ID inválido.")
         with UserData.get_connection() as conn:
             with conn.cursor() as cursor:
                 cursor.execute("SELECT * FROM Usuarios WHERE Cedula = %s;", (cedula,))
                 result = cursor.fetchone()
-                if not result:
-                    raise EntryNotFoundError("Usuario no encontrado.")
+                UserOutput.validate_user_found(result, "consulta")
                 return result
 
+    @staticmethod
+    def user_exists(cedula):
+        with UserData.get_connection() as conn:
+            with conn.cursor() as cursor:
+                cursor.execute("SELECT EXISTS(SELECT 1 FROM Usuarios WHERE Cedula = %s)", (cedula,))
+                return cursor.fetchone()[0]
